# backend/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar Payment API",
    "version": "0.1.0",
    "description": "API for creating and verifying Stellar network payments. Accept Stellar-based payments with simple links and status tracking.",
    "contact": {
      "name": "Stellar Payment API Support",
      "url": "https://github.com/emdevelopa/Stellar_Payment_API"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key for authenticating merchant requests. Obtain this key when registering a merchant."
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/payments/{id}": {
      "get": {
        "summary": "Get detailed payment intent",
        "description": "Returns the full payment record for the authenticated merchant. When status is \"completed\", a `stellar_tx` block is also included containing the on-chain tx_hash, ledger sequence, and timestamp sourced directly from Horizon.\n",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Payment intent UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "number"
                        },
                        "asset": {
                          "type": "string"
                        },
                        "asset_issuer": {
                          "type": "string",
                          "nullable": true
                        },
                        "recipient": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string",
                          "nullable": true
                        },
                        "memo": {
                          "type": "string",
                          "nullable": true
                        },
                        "memo_type": {
                          "type": "string",
                          "nullable": true
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "pending",
                            "confirmed",
                            "completed",
                            "failed"
                          ]
                        },
                        "tx_id": {
                          "type": "string",
                          "nullable": true
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "branding_config": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    },
                    "stellar_tx": {
                      "nullable": true,
                      "type": "object",
                      "description": "Present only when status is \"completed\" and Horizon returns the transaction. null otherwise.\n",
                      "properties": {
                        "tx_hash": {
                          "type": "string"
                        },
                        "ledger": {
                          "type": "integer"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorised — invalid or missing API key"
          },
          "404": {
            "description": "Payment not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Soft delete a payment (preserves audit logs)",
        "tags": [
          "Payments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Payment ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment soft deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "payment_id": {
                      "type": "string"
                    },
                    "deleted_at": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Payment not found"
          },
          "410": {
            "description": "Payment already deleted"
          }
        }
      }
    },
...
    "/api/webhook-logs": {
      "get": {
        "summary": "Get webhook delivery logs for authenticated merchant",
        "description": "Retrieve paginated webhook delivery logs for the authenticated merchant account using cursor-based pagination",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            },
            "description": "Pagination cursor (base64 encoded)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Number of logs per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "success",
                "failure"
              ]
            },
            "description": "Filter by success (2xx) or failure (non-2xx)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated webhook logs"
          },
          "401": {
            "description": "Unauthorized - invalid or missing API key"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/webhooks/test": {
      "post": {
        "summary": "Send a test webhook to the merchant's stored webhook URL",
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/notifications": {
      "get": {
        "summary": "Get dashboard notifications",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  }
}